move `reorderIndexForReorder` to `flex-reorder-strategy`
